src/main/java/seedu/address/logic/commands/ReturnCommand.java

package seedu.address.logic.commands;

import seedu.address.commons.core.Messages;
import seedu.address.commons.core.index.Index;
import seedu.address.logic.commands.exceptions.CommandException;
import seedu.address.model.Model;
import seedu.address.model.book.exceptions.BookNotFoundException;
import seedu.address.model.book.Book;
import seedu.address.model.book.exceptions.DuplicateBookException;

import java.util.List;
import java.util.Objects;

import static java.util.Objects.requireNonNull;

    /**
     * Un-mark a borrowed book to make it available for borrowing
     */

    public class ReturnCommand extends UndoableCommand {

        public static final String COMMAND_WORD = "return";

        public static final String MESSAGE_RETURN_BOOK_SUCCESS = "Book is returned: %1$s";

        public static final String MESSAGE_USAGE = COMMAND_WORD + ": Returns book whose title matches the input \n"
                + "Parameters: INDEX (must be a positive integer) \n"
                + "Example: " + COMMAND_WORD + " Harry Potter and the Half Blood Prince";

        public static final String MESSAGE_BOOK_CANNOT_BE_RETURNED = "Book cannot be returned as it is already available";

        public static final int PRIVILEGE_LEVEL = Model.PRIVILEGE_LEVEL_LIBRARIAN;

        private final Index targetIndex;

        private Book bookToReturn;

        public ReturnCommand(Index targetIndex) {
            this.targetIndex = targetIndex;
        }

        @Override
        public CommandResult executeUndoableCommand() throws CommandException {
            requireNonNull(bookToReturn);
            try {
                model.returnBook(bookToReturn);
                return new CommandResult(String.format(MESSAGE_RETURN_BOOK_SUCCESS, bookToReturn));
            }catch (BookNotFoundException pnfe) {
                throw new CommandException(MESSAGE_BOOK_CANNOT_BE_RETURNED);
            }
        }

        @Override
        protected void preprocessUndoableCommand() throws CommandException {
            List<Book> lastShownList = model.getFilteredBookList();

            if (targetIndex.getZeroBased() >= lastShownList.size()) {
                throw new CommandException(Messages.MESSAGE_INVALID_BOOK_DISPLAYED_INDEX);
            }

            bookToReturn = lastShownList.get(targetIndex.getZeroBased());
        }

        @Override
        public boolean equals(Object other) {
            return other == this // short circuit if same object
                    || (other instanceof ReturnCommand // instanceof handles nulls
                    && this.targetIndex.equals(((ReturnCommand) other).targetIndex) // state check
                    && Objects.equals(this.bookToReturn, ((ReturnCommand) other).bookToReturn));
        }

        @Override
        public int getPrivilegeLevel() {
            return PRIVILEGE_LEVEL;
        }
    }

